Teach GPROCA about the 'anyenum' and 'anynonarray' polymorphic types.
PostgreSQL 8.3 comes with these two new polymorphic types. ORCA needs
to know about them, otherwise the PostgreSQL 'enum' regression test fails
with this error:

ERROR:  type matched to anyenum is not an enum type: anyenum

The knowledge about different GPDB data types would probably belong in
the GPDB repository, so that we wouldn't need to modfiy GPORCA every time
there are changes to the built-in types in GPDB or PostgreSQL, but that's
a different story.