Remove unused support for dynamic min/max abs value values in CDouble.
It was possible to specify a minimum and maximum abs value for a CDouble,
in the constructor. That would allow using different minimum and maximum
in different places. However, that facility was unused; no caller passed
non-default min/max values. So remove the unnecessary flexibility, and
just use the same, default limits everywhere.

I was seeing compiler warnings on failed inlining in GPDB from the CDouble's
CheckValidity function, with -Winline. As well as saving some memory and
some cycles from all operations on CDoubles and making the code simpler,
I'm hoping that this will make those warnings go away.